+++ extensions/project-companion/src/index.tsx
@@ -10,6 +10,7 @@ import {
   confirmAlert,
   Alert,
 } from "@raycast/api";
+import { useForm, FormValidation } from "@raycast/utils";
 import { useState, useEffect } from "react";
 
 interface Project {
@@ -207,7 +208,7 @@ export default function Command() {
 function CreateProjectForm(props: { onCreate: (project: Project) => void }) {
   const { pop } = useNavigation();
 
-  function handleSubmit(values: {
+  const submitForm = (values: {
     title: string;
     status: string;
     backend: string;
@@ -217,7 +218,7 @@ function CreateProjectForm(props: { onCreate: (project: Project) => void }) {
     roadmap: string;
     design: string;
     favorite: string;
-  }) {
+  }) => {
     props.onCreate({
       title: values.title,
       status: values.status,
@@ -230,7 +231,14 @@ function CreateProjectForm(props: { onCreate: (project: Project) => void }) {
       favorite: values.favorite,
     });
     pop();
-  }
+  };
+
+  const { handleSubmit, itemProps } = useForm({
+    onSubmit: submitForm,
+    validation: {
+      title: FormValidation.Required,
+    },
+  });
 
   // Create View
   return (
@@ -241,7 +249,7 @@ function CreateProjectForm(props: { onCreate: (project: Project) => void }) {
         </ActionPanel>
       }
     >
-      <Form.TextField id="title" title="Title" placeholder="project name" />
+      <Form.TextField title="Title" placeholder="project name" {...itemProps.title} />
       <Form.TextArea
         id="description"
         title="Project Description"
@@ -291,7 +299,7 @@ function DeleteProjectAction(props: { onDelete: () => void }) {
     <Action
       icon={Icon.Trash}
       title="Delete Project"
-      shortcut={{ modifiers: ["cmd", "shift"], key: "delete" }}
+      shortcut={{ modifiers: ["ctrl"], key: "delete" }}
       onAction={props.onDelete}
       style={Action.Style.Destructive}
     />
@@ -304,8 +312,11 @@ function EditProjectForm(props: {
   index: number;
 }) {
   const { pop } = useNavigation();
+  const [title, setTitle] = useState(props.project.title);
 
-  function handleSubmit(values: {
+  const initialValues = { title };
+
+  const submitForm = (values: {
     title: string;
     status: string;
     backend: string;
@@ -316,7 +327,7 @@ function EditProjectForm(props: {
     design: string;
     other: string;
     favorite: string;
-  }) {
+  }) => {
     props.onEdit(props.index, {
       title: values.title,
       status: values.status,
@@ -330,9 +341,17 @@ function EditProjectForm(props: {
       favorite: values.favorite,
     });
     pop();
-  }
+  };
 
-  // Create View
+  const { handleSubmit, itemProps } = useForm({
+    onSubmit: submitForm,
+    validation: {
+      title: FormValidation.Required,
+    },
+    initialValues,
+  });
+
+  // Edit View
   return (
     <Form
       actions={
@@ -341,7 +360,14 @@ function EditProjectForm(props: {
         </ActionPanel>
       }
     >
-      <Form.TextField id="title" title="Title" placeholder="project name" defaultValue={props.project.title} />
+      <Form.TextField
+        title="Title"
+        placeholder="project name"
+        value={title}
+        onChange={(newValue) => setTitle(newValue)}
+        {...itemProps.title}
+      />
+
       <Form.Dropdown id="status" title="Status" defaultValue={props.project.status}>
         {projectStatus.map((status, index) => (
           <Form.Dropdown.Item
@@ -387,7 +413,7 @@ function EditProjectAction(props: {
     <Action.Push
       icon={Icon.Pencil}
       title="Edit Project"
-      shortcut={{ modifiers: ["cmd", "shift"], key: "e" }}
+      shortcut={{ modifiers: ["ctrl"], key: "e" }}
       target={<EditProjectForm onEdit={props.onEdit} project={props.project} index={props.index} />}
     />
   );